move tests to their own classes
